Ship Movement
Turning is still hella fucked, but forward and backwards both work. I temporarily created a separate water shader, but it shouldn't be necessary anymore. The big thing that's needed is the Waves script, as the ship gets part of it's movement information from that script (bobbing up and down and such).